LayoutTests/accessibility/color-well.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../resources/js-test.js"></script>
<script src="../resources/accessibility-helper.js"></script>
</head>
<body>

<input id="empty_colorwell" type="color">
<input id="good_colorwell" type="color" value="#ff0000">
<input id="bad_colorwell" type="color" value="purple">
    
<script>
    description("This test checks the role of ColorWellRolean input with type=color");

    if (window.accessibilityController) {
        window.jsTestIsAsync = true;

        setTimeout(async function() {
            let axColorwell = await waitForElementById("empty_colorwell");
            debug("Role of input type=color is: " + axColorwell.role);
            debug("Value of empty color well: " + axColorwell.stringValue);

            axColorwell = await waitForElementById("good_colorwell");
            debug("Value of good color well: " + axColorwell.stringValue);

            axColorwell = await waitForElementById("bad_colorwell");
            debug("Value of bad color well: " + axColorwell.stringValue);

            finishJSTest();
        }, 0);
    }
</script>
</body>
</html>
