Safari-611.1.10/LayoutTests/accessibility/color-well-expected.txt


This test checks the role of ColorWellRolean input with type=color

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Role of input type=color is: AXRole: AXColorWell
Value of empty color well: AXValue: rgb 0.00000 0.00000 0.00000 1
Value of good color well: AXValue: rgb 1.00000 0.00000 0.00000 1
Value of bad color well: AXValue: rgb 0.00000 0.00000 0.00000 1
PASS successfullyParsed is true

TEST COMPLETE
